requirements.txt
acres==0.2.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
async-timeout==4.0.3
attrs==24.3.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
comm==0.2.2
configobj==5.0.9
configparser==7.1.0
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
distro==1.9.0
etelemetry==0.3.1
exceptiongroup==1.2.2
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.17.0
fitz==0.0.1.dev2
fonttools==4.55.4
fqdn==1.5.1
frontend==0.0.3
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.31.0
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_geojson==3.4.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.8
langchain==0.3.15
langchain-core==0.3.31
langchain-openai==0.3.2
langchain-text-splitters==0.3.5
langsmith==0.1.147
llmwhisperer-client==2.0.0
looseversion==1.3.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.0
multidict==6.1.0
narwhals==1.25.2
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.2
notebook_shim==0.2.4
numpy==1.26.4
openai==1.60.1
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathlib==1.0.1
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.5
pexpect==4.9.0
pikepdf==9.5.1
pillow==11.1.0
platformdirs==4.3.6
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==5.29.3
prov==2.0.1
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.28
pyarrow==19.0.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pydot==3.0.4
Pygments==2.19.1
PyMuPDF==1.25.2
pyparsing==3.2.1
pypdfium2==4.30.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
pytz==2024.2
pyxnat==1.6.2
PyYAML==6.0.2
pyzmq==26.2.0
rdflib==6.3.2
referencing==0.36.1
regex==2024.11.6
reportlab==4.2.5
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
scipy==1.15.1
Send2Trash==1.8.3
simplejson==3.19.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
stack-data==0.6.3
starlette==0.45.2
streamlit==1.42.0
tenacity==8.5.0
terminado==0.18.1
tiktoken==0.8.0
tinycss2==1.4.0
toml==0.10.2
tomli==2.2.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
traits==6.4.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
wrapt==1.17.2
yarl==1.18.3
